requirements.txt
selenium
watchdog
pillow
cv2
requests
python-dotenv
